unify batch code
